pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepm"
version = "1.0.0"
description = "Deep Momentum Networks with Bipartite Graph Attention"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [
    {name = "Kieran Wood"},
]

dependencies = [
    "torch>=2.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "wandb>=0.15",
    "colorama>=0.4",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    "pyarrow>=14.0",
    "joblib>=1.3",
]

[project.urls]
Paper = "https://arxiv.org/abs/2601.05975"

[tool.setuptools.packages.find]
include = ["deepm*"]

[tool.setuptools.package-data]
deepm = ["configs/**"]
